Modify the TokenGenerator to monitor the signing certificate:
Add a timer to the constructor that runs every five minutes and checks to see if the
modulus of the existing in-memory publicKey equals the current server public key's
modulus.  If not, set the in-memory public key to the new one (as well as the 
private key and consumer key).