Update the TokenGenerator to generate multiple long term tokens:
This change allows the main() method to now take three arguments:

- The path to a mounted Keybase team to ensure Keybase directories are writable
- The path to the JSON configuration file with token subject details
- The path to the private certificates and keys directory used for signing

The JSON file is parsed and a token is created and signed for each listed identity. The tokens are then written to the Keybase directory stated in the configuration file for the given identity.

Also include a method called getFutureShortLivedJWT() which will create a token that will be valid in the future and will shortly expire.  This is used for testing where we need to pre-generate a number of tokens.